MDL-46951 gradereport_grader: Handling of user cells when resizing
For some reason, the max-width/min-width/width of the table cell containing
the name of the users did not work, leading to very large columns but a very
narrow floating one. Handling the size of those columsn upon resize was
also necessary after with this fix.

This issue is a part of the MDL-46658 Task.
This issue is a part of the MDL-25544 Epic.