MDL-46936 gradereport_grader: AJAX fixes for gradereport
Fixed functionality of arrow-key movement on all browsers.
Added in ESC key during non-edit mode to revert any changes to a cell.
Fixed positioning of analysis icon in cells being edited.
Corrected AJAX not loading problem, id no longer supplied to module.js.

This issue is a part of the MDL-46658 Task.
This issue is a part of the MDL-25544 Epic.